CFAM: Define CFAM reg fields and extraction
The CFAM-S Gemini mailbox scratchpad registers are broken up into
various fields to hold redundant BMC information.  Fields can be from
1 to 32 bits in length.

This commit defines the fields. This includes the name, register, start
bit, and length. These are stored in a 'BMCCFAM' base class. There are
also new functions to extract the fields out of uint32_t register
values.

When the registers on the sibling BMC are polled, all registers will be
read at once, so there is also a function to extract a field out of a
map of register name to values.

When a field is modified, under the covers there will have to be a read,
followed by placing the field into the proper offset without writing
over other fields, followed by a write.  To handle this, there is a new
createModifyOp() function that will create a ModifyOp structure to hold
the field value shifted into the proper location along with the mask of
the field so existing bits in the register can be left alone.

The plan is to derive LocalCFAM and SiblingCFAM child classes from the
base class, and in those classes make calls to extract fields after the
raw data is read.

Tested:
New unit tests pass.

Change-Id: Ia81a85e29ec4a0a367d943c93c44ddf209076fb8
Signed-off-by: Matt Spinler <spinler@us.ibm.com>