CFAM: APIs to read/write the CFAM
This commit creates a CFAMAccess class that has methods for reading and
writing the CFAM scratchpad registers.  There is also a SysFS base class
along with its implementation class to do the reading and writing of the
sysfs files so it can be mocked.

The FSI sysfs driver FSI is such that to do a write one just writes the
raw 4 bytes into the sysfs file, but when reading the file it returns an
8 byte string representation of the value:

```
$ hexdump -C scratch1
00000000  31 32 33 34 35 36 37 38   |12345678|
```

Tested:
- Unit tests work.
- Reads and writes work in simulation.

Change-Id: I2bba5928f3c5308a48446503720bbab89de8002e
Signed-off-by: Matt Spinler <spinler@us.ibm.com>