CFAM: Create SiblingCFAM class
This class handles reading fields from the sibling BMC's CFAM scratch
pad registers. It knows where they are in the registers because it
inherits from the BMCCFAM class.

It provides a single method readAll() that will read all registers that
have fields in them and then load the field values into variables
representing the fields, which there are then access functions for.

The idea for use is that readAll() will be called periodically to
obtain the sibling's redundancy related values.

If a register access fails, the 'error' member will be set in the class
and all get functions will throw an exception.  It will stay this way
until there readAll() runs cleanly again.  It is all or nothing, either
one can get all values, or none.

Tested:
Works as expected in the simulator.

Change-Id: Ibf7effe75f3c42386d8a2f441e888892e78eaa9d
Signed-off-by: Matt Spinler <spinler@us.ibm.com>