CFAM: Create Services class to gather data
This new Services class has APIs to read the following fields from the
system:
* BMC state
* Redundancy enabled
* Redundancy role
* FW version ID from /etc/os-release
* BMC position

They need to be obtained so they can be written into the local BMC's
CFAM.

The first 3 fields are on D-Bus, so the constructor takes callback
functions that will be called with the new value when the values change
on D-Bus (either PropertiesChanged or InterfacesAdded).

The FW Version field is read straight out of /etc/os-release and not
from anywhere on D-Bus so it is simpler - i.e. no need to track down
associations to find the active image nor have to wait for it to even
show up on D-Bus.

The method to obtain the BMC position is just a temporary one that works
in the current simulation model until the actual code that reads it from
FSI cables is available.  The simics framework writes the position into
the 'bmc_position' uboot environment variable.

Change-Id: I2ce0f685b4f9bed6869c6952e1c671f1281c2345
Signed-off-by: Matt Spinler <spinler@us.ibm.com>