CFAM: Start populating the local CFAM's regs
This commits creates a LocalBMC class to handle populating the local
BMC's CFAM register fields.  That class is added to the Application
class, where it calls the LocalBMC::start() method near the beginning of
its own run() method.

LocalBMC gathers its data using the Services class, which it provides
callback functions to so it can handle the D-Bus properties like the BMC
state, role, and redundancy enabled changing after startup.

It also watches for the Heartbeat signal on the redundancy D-Bus
interface, and will increment the heartbeat field in the CFAM on every
signal.  The code running on the sibling would then read that to know
this BMC is alive.

For now, the code will wait up to 5 minutes for the CFAM's sysfs files
to show up, and then just give up and crash if they aren't there.  This
area needs more thought, as we don't want it to just continue running
such that the BMC becomes active, but if it crashes and restarts every
10 minutes it probably won't be able to quiesce the BMC, which would be
ideal.

After more things get further along, like the actual system design and
the device drivers to go along with the FSI FPGA, this can be revisited
to see if there's a better way to handle.  It should be a good enough
starting point though so that we're not holding up all of the code that
needs the CFAM's written just for this error path.

Tested:
Using cfamtool, can see the BMC fields are written to the CFAM, and
change when the corresponding D-Bus properties change.

The heartbeat field can be seen to be incremented to match the signal.

Change-Id: Ic2d6b187a9be9f1a57fc53c8aaa11d782d73337e
Signed-off-by: Matt Spinler <spinler@us.ibm.com>