CFAM: Create redundant BMC CFAM-S app framework
Some background:
Each BMC in IBM's redundant BMC system has an FSI connection to its own
CFAM-S (CFAM-S = a standalone CFAM) as well as the sibling BMC's CFAM-S.
The four scratchpad registers in the Gemini mailbox block in these CFAMs
will be used for communication between the BMCs.

The local BMC will write the registers in its own CFAM.  The sibling BMC
would then read these to obtain the information.

On startup the local BMC will collect the information the sibling needs
and write it into its CFAM.  It may also watch D-Bus in case it needs to
update a field on a change.

The sibling's CFAM will be read every N seconds, where N = 2 in this
commit to start with. A D-Bus interface will be created to make these
values available to other applications.

This commit creates a new application rbmc-cfamd to do this.  It is
enabled with a new 'rbmc-cfam-daemon' meson option.  Initially it will
just go into a 2 second loop, not doing anything.

The commit doesn't include a service file because it will be named after
its D-Bus service name, which in turn is named after it's D-Bus
interface, and that isn't in yet.

Change-Id: I1fcd62cc550f0f78bd4be1852417ad7124ef664c

Change-Id: Idd92c05dc9f99fc0b30efef94d307b04352f2084
Signed-off-by: Matt Spinler <spinler@us.ibm.com>