CFAM: Support new device driver
There is a new version of the FSI/CFAM device driver. The changes are:
- It has ioctl() calls for read/write/writeWithMask which take the
  register as a number between 0 and 3 and a descriptor for the CFAM
  device in /dev.
- It requires use of a new fsi.h header file which isn't in the kernel
  yet so it's included locally.

The Driver class is updated accordingly. It looks up the /dev file to
use by first looking it up in /sys/class/fsi-master/ based on the FSI
link number.

So that unit tests can still work, there are new functions to return the
/dev and /sys/class/fsi-master prefixes so that in test a different
version can run that returns temporary directories.

Tested:
- Everything still works in simulation.

Change-Id: I745791937b60a92941f6b30ce0e500c0c1b977cc
Signed-off-by: Matt Spinler <spinler@us.ibm.com>